chore: Configure jest to use a tsconfig that preserves jsx
For next, we use jsx: preserve but jest needs normal javascript to
be able to run the tests. This is why we use a special tsconfig.test.json
that overrides the default jsx attribute, when we run the tests. This
means that we have to be careful to launch jest from the root directory
when working on tests; otherwise, the right jest.config is not selected,
thus the right tsconfig is not selected (😅).