fix: Ability to override chart config filter with "None" in interactive filters
Instead of deleting the value, we have to keep the none value in the
interactive filter state, otherwise the value taken is from configurator
state = the user selects "None" and the state of the select goes back
to the value from configurator state; and the chart data is still
filtered.
Since None is kept inside the state, we need to filter it before sending
the actual query.