fix: Ability to fetch dimension values when datatypes has undefined values
When the dataset has undefined values, the shacl constraint for dataType
is expressed as an "or" containing two constraints (undefined, and the
real datatype). The parseDataType function handles this case and provides
us with the correct datatype.
Shared dimensions have no dataType and the filter for values will be
done with a named node ?dimension1 = <value>, whereas for literal dimensions,
we use the str function to match on the string value of the value
(discarding the type information)