feat: Do not use router to save URL as replaceState has too much functionality baked in
router.replace would cause problems since it was called with all the
previous state, including the href. Since the href could have unbaked
params ("/browse/[type]/[iri]"), we would have had to call it
with the variables necessary for router to interpolate the full path.

This caused the following error on the /en/browse/dataset/<iri> route:

```
The provided href value is missing query values to be interpolated properly
```

Since our only goal with the dataSource storage is to save the param
for URL sharing, we can use the simpler history.replaceState API