fix: Passing chartConfig down to sorting functions so that only filtered values are considered
Previously, when sorting values, we have to build a dimension value
index, indexed by label, since the data that we receive does not contain
the value iris, but it contains the value labels.

The problem occurs
when two dimension values have the same label, but not the same IRI,
for example in the case of hierarchy, where two different regions
are named the same but are located in two different parts of the tree.

Previously, the sorting would end up wrong, since we would make a mistake
when getting the dimension value (finding one dimension value from the
index, when we should have found the other).

To remediate the problem, when there is a filter on the dimension, we
pass it to the function creating the sorting, so that only values
present in the filter are used. This way, we lower the risk of using
two values with the same label.

The problem would still be there if the chart creator selects two
values with the same label, to be displayed at the same time on the
same chart. For example if someone decides to display at the same time
Bern the canton, and Bern the city. Then we would have trouble indexing
correctly. In this case, the person reading the chart would also have
trouble though.